moved GlobalStyle component to styles directory
